add zod validation on server to delete a post on db

+++ src/actions/actions.ts
@@ -39,19 +39,19 @@ export async function addPost(newPost: unknown) {
   revalidatePath("/feed", "layout")
 }
 
-const selectedPostIdSchema = z.number()
+const postIdSchema = z.number()
 
-export async function editPost(editedData: unknown, selectedPostId: unknown) {
+export async function editPost(editedData: unknown, postId: unknown) {
   const failMessage = { message: "Could not edit post. Please, try again in a few minutes." }
   try {
     await delay(1000)
     const validatedEditedData = postFormSchema.safeParse(editedData)
-    const validatedSelectedPostId = selectedPostIdSchema.safeParse(selectedPostId)
-    if (!validatedEditedData.success || !validatedSelectedPostId.success) {
+    const validatedPostId = postIdSchema.safeParse(postId)
+    if (!validatedEditedData.success || !validatedPostId.success) {
       return { message: "Invalid edited post data." }
     }
 
-    const response = await fetch(`${baseUrl}${validatedSelectedPostId.data}/`, {
+    const response = await fetch(`${baseUrl}${validatedPostId.data}/`, {
       method: "PATCH",
       headers: { "content-type": "application/json" },
       body: JSON.stringify(validatedEditedData.data)
@@ -71,7 +71,12 @@ export async function deletePost(postId: unknown) {
   const failMessage = { message: "Could not delete post. Please, try again in a few minutes." }
   try {
     await delay(1000)
-    const response = await fetch(`${baseUrl}${postId}/`, {
+    const validatedPostId = postIdSchema.safeParse(postId)
+    if (!validatedPostId.success) {
+      return { message: "Invalid post id." }
+    }
+
+    const response = await fetch(`${baseUrl}${validatedPostId.data}/`, {
       method: "DELETE",
       headers: { "content-type": "application/json" },
       body: JSON.stringify({})
